Many miscellaneous changes. was forgetting to commit because chaning directly on master computer. This includes changing UINT to unsigned, which is the cause for almost everything to have small changes here.

+++ Chimera/Source/AnalogInput/AiSettings.h
@@ -3,7 +3,7 @@
 struct AiSettings
 {
 	bool queryContinuously=false;
-	double continuousModeInterval=1000;
+	int continuousModeInterval=1000;
 	bool queryBtwnVariations=false;
 	unsigned int numberMeasurementsToAverage=100;
 };
